requirements.txt
Django==3.2.9
whitenoise==5.3.0
gunicorn==20.1.0
psycopg2==2.9.2
black==22.3.0
pre-commit==2.12.0
python-dotenv==0.19.2
htmldocx~=0.0.6
docxtpl==0.15.2
django-nested-admin==3.4.0
Pillow==9.2.0
fixtures==4.0.1
pytest==7.2.0
pytest-django==4.5.2
unittest2==1.1.0
aspose-words==22.11.0